✨ Feat: Map Api,Dog Api 구현완료

+++ server/services/mapService.js
@@ -1,9 +1,12 @@
 const Map = require("../models/mapModel.js");
-
 const mapService = {
 	// 마커 데이터 생성
-	async createMap(mapData) {
+	async createMap(mapData, currentUserId) {
 		try {
+			if (!currentUserId) {
+				throw { status: 400, message: "사용자 ID가 제공되지 않았습니다." };
+			}
+			mapData.userId = currentUserId;
 			const mapProfile = await Map.create(mapData);
 			const mapObject = mapProfile.toObject();
 			return mapObject;
@@ -12,13 +15,15 @@ const mapService = {
 		}
 	},
 	// 마커 정보 수정
-	async updatedMapProfile(id, mapData) {
+	async updatedMapProfile(id, mapData, currentUserId) {
 		try {
 			const mapProfile = await Map.findById(id).lean();
 			if (!mapProfile) {
 				throw { status: 404, message: "해당 맵 데이터가 존재하지 않습니다." };
 			}
-
+			if (mapProfile.userId !== currentUserId) {
+				throw { status: 403, message: "해당 사용자에게 삭제 권한이 없습니다." };
+			}
 			// 업데이트 수행
 			const updateResult = await Map.updateOne({ _id: id }, mapData).lean();
 
@@ -36,13 +41,16 @@ const mapService = {
 		}
 	},
 
-	// 특정 마커 삭제
-	async deleteMap(id) {
+	async deleteMap(id, currentUserId) {
 		try {
-			const deleteMap = await Map.findByIdAndDelete(id);
-			if (!deleteMap) {
+			const mapToDelete = await Map.findById(id);
+			if (!mapToDelete) {
 				throw { status: 404, message: "해당 맵 데이터가 존재하지 않습니다." };
 			}
+			if (mapToDelete.userId !== currentUserId) {
+				throw { status: 403, message: "해당 사용자에게 삭제 권한이 없습니다." };
+			}
+			const deleteMap = await Map.findByIdAndDelete(id);
 			return deleteMap;
 		} catch (error) {
 			throw error;
@@ -79,11 +87,9 @@ const mapService = {
 
 	async getMapsByTag(tagName) {
 		try {
-			// 태그 형식을 검증
 			if (!isValidTag(tagName)) {
 				throw { status: 400, message: "올바른 태그 형식이 아닙니다." };
 			}
-
 			const mapsByTag = await Map.find({ tag: tagName }).lean();
 			if (!mapsByTag || mapsByTag.length === 0) {
 				throw {
@@ -96,7 +102,22 @@ const mapService = {
 			throw error;
 		}
 	},
+	async getMyMaps(currentUserId) {
+		try {
+			const mapsByUser = await Map.find({ userId: currentUserId }).lean();
+			if (!mapsByUser || mapsByUser.length === 0) {
+				throw {
+					status: 404,
+					message: `해당 사용자의 맵 데이터가 존재하지 않습니다.`,
+				};
+			}
+			return mapsByUser;
+		} catch (error) {
+			throw error;
+		}
+	},
 };
+
 function isValidTag(tagName) {
 	return typeof tagName === "string" && tagName.length > 0;
 }
